pyproject.toml
[project]
name = "broverse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "broflow>=0.1.0",
    "brollm==0.1.1",
    "graphviz>=0.21",
    "matplotlib>=3.10.3",
    "mermaid-py>=0.8.0",
    "networkx>=3.5",
    "plotly>=6.2.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "boto3>=1.39.14",
    "ipykernel>=6.30.0",
    "ipywidgets>=8.0.0",
    "jupyter>=1.0.0",
    "requests>=2.32.4",
    "sentence-transformers>=5.0.0",
    "tokenizers==0.19.1",
    "transformers>=4.44.2",
]

[tool.setuptools.packages.find]
include = ["broverse*"]
